add some more selenium tests for ASP.NET
