+++ samples/web/CQELight_BlazorServer/Pages/Index.razor
@@ -2,6 +2,6 @@
 
 <h1>Hello, world!</h1>
 
-Welcome to your new app.
+<p>If you see this, it means that your Blazor website is running with CQELight ! </p>
 
 <SurveyPrompt Title="How is Blazor working for you?" />
