+++ samples/web/CQELight_BlazorServer/Pages/_Host.cshtml
@@ -10,7 +10,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>CQELight_BlazorServer</title>
+    <title>CQELight demo website - Hybrid Technologies Solutions</title>
     <base href="~/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
     <link href="css/site.css" rel="stylesheet" />
